move custom types to into declaration file
