+++ tsconfig.json
@@ -26,7 +26,7 @@
       "@customTypes/*": ["./customTypes/*"],
       "@utils/*": ["./utils/*"]
     },
-    "typeRoots": ["node_modules/@types", "./types"]
+    "typeRoots": ["node_modules/@types", "./types", "./src/@types"]
   },
   "include": [
     "src/**/*.ts",
